chore: Update GIS Template App configuration
The appConfig.json file was modified to update the titles and URLs of the feature layers. The "Hazardous Materials Facilities" layer was changed to "Tourist Attractions" with a new URL, and the "Parcels" layer was changed to "Beer-Places" with a new URL. The view orders and symbology for both layers were also updated.

Note: This commit message follows the conventional format of "chore: <description>".